add 3-mul.c

+++ 0x0A-argc_argv/3-mul.c
@@ -9,7 +9,7 @@
 int main(int argc, char *argv[])
 {
 	int i, value;
-	long int prod;
+	unsigned int prod;
 
 	if (argc > 1)
 	{
